pyproject.toml
[project]
name = "rasc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.8.5",
    "aiohue==4.7.0",
    "aiosignal==1.3.1",
    "async-timeout==4.0.3",
    "asyncio-throttle==1.0.2",
    "attrs==23.1.0",
    "autograd==1.6.2",
    "awesomeversion==23.11.0",
    "contourpy==1.2.0",
    "cycler==0.12.1",
    "fonttools==4.46.0",
    "frozenlist==1.4.0",
    "future==0.18.3",
    "idna==3.6",
    "kiwisolver==1.4.5",
    "matplotlib==3.8.2",
    "mpmath==1.3.0",
    "multidict==6.0.4",
    "numdifftools==0.9.41",
    "numpy==1.26.2",
    "packaging==23.2",
    "pillow==10.1.0",
    "pyparsing==3.1.1",
    "python-dateutil==2.8.2",
    "pyyaml==6.0.1",
    "scipy==1.11.4",
    "six==1.16.0",
    "sympy==1.12",
    "yarl==1.9.4",
]
